+++ src/api/validations/user.validation.ts
@@ -1,19 +1,13 @@
 export {};
 import * as Joi from 'joi';
-import { User } from 'api/models';
+import { User } from '../../api/models';
 
-const requireEmail = () =>
-  Joi.string()
-    .email()
-    .required();
+const requireEmail = () => Joi.string().email().required();
 
 const postPutBody = () => {
   return {
     email: requireEmail(),
-    password: Joi.string()
-      .min(6)
-      .max(128)
-      .required(),
+    password: Joi.string().min(6).max(128).required(),
     name: Joi.string().max(128),
     role: Joi.string().valid(User.roles)
   };
@@ -24,9 +18,7 @@ module.exports = {
   listUsers: {
     query: {
       page: Joi.number().min(1),
-      perPage: Joi.number()
-        .min(1)
-        .max(100),
+      perPage: Joi.number().min(1).max(100),
       name: Joi.string(),
       email: Joi.string(),
       role: Joi.string().valid(User.roles)
@@ -52,9 +44,7 @@ module.exports = {
   updateUser: {
     body: {
       email: Joi.string().email(),
-      password: Joi.string()
-        .min(6)
-        .max(128),
+      password: Joi.string().min(6).max(128),
       name: Joi.string().max(128),
       role: Joi.string().valid(User.roles)
     },