+++ src/config/passport.ts
@@ -4,7 +4,7 @@ const bearerStrategy = require('passport-http-bearer');
 const { ExtractJwt } = require('passport-jwt');
 const { JWT_SECRET } = require('./vars');
 const authProviders = require('../api/services/authProviders');
-import { User } from 'api/models';
+import { User } from '../api/models';
 
 const jwtOptions = {
   secretOrKey: JWT_SECRET,