+++ src/api/routes/v1/auth.route.ts
@@ -4,6 +4,7 @@ const validate = require('express-validation');
 const controller = require('../../controllers/auth.controller');
 const oAuthLogin = require('../../middlewares/auth').oAuth;
 const { login, register, oAuth, refresh, forgotPassword } = require('../../validations/auth.validation');
+const { authorize, LOGGED_USER } = require('../../middlewares/auth');
 
 const router = express.Router();
 
@@ -131,5 +132,6 @@ router.route('/facebook').post(validate(oAuth), oAuthLogin('facebook'), controll
 router.route('/google').post(validate(oAuth), oAuthLogin('google'), controller.oAuth);
 
 router.route('/forgot-password').post(validate(forgotPassword), controller.forgotPassword);
+router.route('/logout').post(authorize(LOGGED_USER), controller.logout);
 
 module.exports = router;