+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.10.0] - 2020-11-05
 - upgraded dependencies.
 - added: endpoint: create User Note: POST /users/USERID/notes - payload { title, note }
+- added a CRA v4 webapp as an example to access APIs.
 
 ## [0.9.0] - 2019-10-27